requirements.txt
pandas>=2.0.0
selenium>=4.0.0
webdriver-manager>=4.0.0
sqlite3-binary; python_version < "3.12" 
apache-airflow==3.1.3